requirements.txt
fastapi==0.115.0
nest_asyncio==1.6.0
osu.py==2.2.1
python-dotenv==1.0.1
uvicorn==0.31.0
